Backend/summarizer/llm.py
import os
from langchain_groq import ChatGroq
from dotenv import load_dotenv

load_dotenv()
def get_llm():
    api_key = os.getenv("GROQ_API_KEY")

    if not api_key:
        raise ValueError("GROQ_API_KEY not found in environment variables")
    return ChatGroq(
        groq_api_key=api_key,
        model_name="openai/gpt-oss-120b",
        temperature=0,
    )