Backend/requirements.txt
asgiref
Django
djangorestframework
sqlparse
tzdata

langchain
langchain-community
groq
python-dotenv
langchain_groq
